+++ src/main/java/net/oneandone/maven/plugins/dockerbuild/Build.java
@@ -36,7 +36,6 @@ import java.io.InputStream;
 import java.io.PrintWriter;
 import java.net.InetAddress;
 import java.net.UnknownHostException;
-import java.util.ArrayList;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
@@ -162,8 +161,6 @@ public class Build extends Base {
         project.getProperties().put("dockerbuild.origin", origin());
         Arguments a = context.arguments(log);
         a.addArtifacts(context, world.file(project.getBuild().getDirectory()), artifactName);
-        a.addFiles(srcdirs(), buildDirectory(), fileFilter, project, session);
-        a.addPom(project);
         a.addExplicit(arguments, world, fileFilter, project, session);
         actuals = a.result();
         try (InputStream tarSrc = context.tar().newInputStream()) {
@@ -187,16 +184,6 @@ public class Build extends Base {
         }
         log.info("Done: " + repositoryTag);
         log.debug("id=" + id + " seconds=" + (System.currentTimeMillis() - started) / 1000);
-    }
-
-    private List<FileNode> srcdirs() {
-        List<FileNode> result;
-
-        // CAUTION: ordering is imported, later items precedence
-        result = new ArrayList<>();
-        result.add(buildDirectory().getParent().join("dockerbuildfiles"));
-        result.add(world.file(project.getBasedir()).join("src/dockerbuild"));
-        return result;
     }
 
     private static String origin() {