+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 ## 1.2.0 (pending)
 
 * renamed `master` branch to `main`
-* dumped `property` arguments - use explicit argument that reference the property instead
+* stop pulling arguments, that's like using global variables - at first, it's handy, but as things grow, it becomes scary magic
+  * dumped `property` arguments - use explicit argument that reference the property instead
 * file arguments now also search in target/dockerbuildfiles for source files  
 * skippable builds
   * build goal: