+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * renamed `master` branch to `main`
 * stop pulling arguments, that's like using global variables - at first, it's handy, but as things grow, it becomes scary magic
+  * added %base64 directive to base64-encode a string
   * dumped `property` arguments - use explicit argument that reference the property instead
 * file arguments now also search in target/dockerbuildfiles for source files  
 * skippable builds